Apply SOLID principles to project_io.py and forge.py
