+++ fastapi_forge/core/project_validators/protocols.py
@@ -1,0 +1,10 @@
+from abc import abstractmethod
+from typing import Protocol
+
+from fastapi_forge.schemas import ProjectSpec
+
+
+class ProjectValidator(Protocol):
+    @abstractmethod
+    def validate(self, project_spec: ProjectSpec) -> None:
+        raise NotImplementedError
