+++ fastapi_forge/core/template_processors/__init__.py
@@ -1,0 +1,7 @@
+__all__ = [
+    "DefaultTemplateProcessor",
+    "TemplateProcessor",
+]
+
+from .processors import DefaultTemplateProcessor
+from .protocols import TemplateProcessor
