+++ fastapi_forge/core/template_processors/protocols.py
@@ -1,0 +1,10 @@
+from abc import abstractmethod
+from typing import Any, Protocol
+
+from fastapi_forge.schemas import ProjectSpec
+
+
+class TemplateProcessor(Protocol):
+    @abstractmethod
+    def process(self, spec: ProjectSpec) -> dict[str, Any]:
+        raise NotImplementedError
