+++ fastapi_forge/io/builder/protocols.py
@@ -1,0 +1,8 @@
+from abc import abstractmethod
+from typing import Protocol
+
+
+class ProjectBuilder(Protocol):
+    @abstractmethod
+    async def build_artifacts(self) -> None:
+        raise NotImplementedError
