+++ fastapi_forge/io/database/__init__.py
@@ -1,0 +1,10 @@
+__all__ = [
+    "DatabaseInspector",
+    "PostgresInspector",
+    "SchemaInspectionResult",
+    "SchemaInspector",
+]
+
+from .postgres_inspector import PostgresInspector
+from .protocols import DatabaseInspector
+from .schema import SchemaInspectionResult, SchemaInspector
