+++ fastapi_forge/io/loader/protocols.py
@@ -1,0 +1,10 @@
+from abc import abstractmethod
+from typing import Protocol
+
+from fastapi_forge.schemas import ProjectSpec
+
+
+class ProjectLoader(Protocol):
+    @abstractmethod
+    def load(self) -> ProjectSpec:
+        raise NotImplementedError
