+++ fastapi_forge/render/engines/protocols.py
@@ -1,22 +1,17 @@
-from abc import ABC, abstractmethod
+from abc import abstractmethod
 from collections.abc import Callable
-from typing import Any
+from typing import Any, Protocol
 
 
-class BaseTemplateEngine(ABC):
-    """Abstract base class for all template engines."""
-
+class TemplateEngine(Protocol):
     @abstractmethod
     def add_filter(self, name: str, filter_func: Callable[[Any], Any]) -> None:
-        """Register a new template filter."""
         raise NotImplementedError
 
     @abstractmethod
     def add_global(self, name: str, value: Any) -> None:
-        """Register a new global variable."""
         raise NotImplementedError
 
     @abstractmethod
     def render(self, template: str, context: dict[str, Any]) -> str:
-        """Render template with given context."""
         raise NotImplementedError
