+++ fastapi_forge/render/manager.py
@@ -1,19 +1,19 @@
-from .engines import BaseTemplateEngine
+from .engines import TemplateEngine
 from .factories import RendererFactory
-from .renderers.base_renderer import BaseRenderer
+from .renderers.protocols import Renderer
 
 
 class RenderManager:
     def __init__(
         self,
-        engine: BaseTemplateEngine,
+        engine: TemplateEngine,
         factories: dict[str, type[RendererFactory]],
     ):
         self.engine = engine
         self.factories = factories
-        self._renderers: dict[str, BaseRenderer] = {}
+        self._renderers: dict[str, Renderer] = {}
 
-    def get_renderer(self, renderer_type: str) -> BaseRenderer:
+    def get_renderer(self, renderer_type: str) -> Renderer:
         """Get a renderer instance for the specified type."""
         if renderer_type not in self.factories:
             raise ValueError(
