+++ fastapi_forge/render/renderers/protocols.py
@@ -1,12 +1,12 @@
-from abc import ABC, abstractmethod
-from typing import Any
+from abc import abstractmethod
+from typing import Any, Protocol
 
 from fastapi_forge.schemas import CustomEnum, Model
 
 Renderable = Model | list[CustomEnum]
 
 
-class BaseRenderer(ABC):
+class Renderer(Protocol):
     @abstractmethod
     def render(self, data: Renderable, **kwargs: Any) -> str:
         raise NotImplementedError
